+++ demos/jax/elastoplasticity/plane_elastoplasticity.md
@@ -34,7 +34,7 @@ from mpi4py import MPI
 import ufl
 from dolfinx import io, fem
 from dolfinx.common import timing, list_timings
-from dolfinx_materials.material import JAXMaterial
+from dolfinx_materials.jaxmat import JAXMaterial
 from dolfinx_materials.quadrature_map import QuadratureMap
 from dolfinx_materials.solvers import NonlinearMaterialProblem
