+++ demos/mfront/heat_transfer/nonlinear_heat_transfer.py
@@ -205,7 +205,7 @@ from dolfinx import fem, mesh
 from dolfinx.cpp.nls.petsc import NewtonSolver
 from dolfinx_materials.quadrature_map import QuadratureMap
 from dolfinx_materials.solvers import NonlinearMaterialProblem
-from dolfinx_materials.material.mfront import MFrontMaterial
+from dolfinx_materials.mfront import MFrontMaterial
 import os
