+++ demos/mfront/heat_transfer/phase_change.md
@@ -212,7 +212,7 @@ from mpi4py import MPI
 from dolfinx import fem, mesh
 from dolfinx_materials.quadrature_map import QuadratureMap
 from dolfinx_materials.solvers import NonlinearMaterialProblem
-from dolfinx_materials.material.mfront import MFrontMaterial
+from dolfinx_materials.mfront import MFrontMaterial
 
 current_path = os.getcwd()
