--- dolfinx_materials/material/__init__.py
@@ -1,19 +1,0 @@
-#!/usr/bin/env python
-# -*-coding:utf-8 -*-
-"""
-
-@Author  :   Jeremy Bleyer, Ecole Nationale des Ponts et Chaussées, Navier
-@Contact :   jeremy.bleyer@enpc.fr
-@Time    :   08/06/2023
-"""
-from .generic import Material
-
-try:
-    from .mfront import MFrontMaterial
-except ImportError:
-    print("MGIS is not available. MFront behaviors cannot be used.")
-
-try:
-    from .jaxmat import JAXMaterial
-except ImportError:
-    print("jax or jaxmat is not available. MFront behaviors cannot be used.")
