+++ tests/mfront/test_nonlinear_elasticity.py
@@ -2,7 +2,7 @@ import pathlib
 import numpy as np
 import sys
 
-from dolfinx_materials.material.mfront import MFrontMaterial
+from dolfinx_materials.mfront import MFrontMaterial
 
 path = pathlib.Path(__file__).parent.absolute()
